+++ src/main/java/com/timeToast/timeToast/dto/payment/PaymentSuccessRequest.java
@@ -3,6 +3,6 @@ package com.timeToast.timeToast.dto.payment;
 public record PaymentSuccessRequest(
         String paymentKey,
         String orderId,
-        long amount
+        int amount
 ) {
 }
